:recycle: refactor editors (not tested yet)
