+++ src/components/common/Editor/RemarkInput.tsx
@@ -7,7 +7,7 @@ import {
   Wrap,
 } from "@chakra-ui/react"
 import { useFormContext } from "react-hook-form"
-import { EditorLabel } from "./CommonEditorComponent"
+import { EditorLabel } from "./CommonEditor"
 
 type RemarkInputProps = {
   label: string
